Fix review/setting/users pages and merge from master
