+++ ui/src/components/flatpicker/FlatPicker.js
@@ -53,9 +53,6 @@ export default class FlatPicker extends React.Component {
 
     onChange(selectedDates, dateStr, instance) {
         this.onChangeDate = dateStr;
-        if (dateStr === '') {
-            this.props.onChange(selectedDates);
-        }
     }
 
     onClose(selectedDates, dateStr, instance) {
