+++ ui/src/components/review/ReviewRow.js
@@ -14,11 +14,9 @@
  * limitations under the License.
  */
 import React from 'react';
-import RadioButton from '../denali/RadioButton';
 import styled from '@emotion/styled';
-import UpdateModal from '../modal/UpdateModal';
+import RadioButton from '../denali/RadioButton';
 import DateUtils from '../utils/DateUtils';
-import RequestUtils from '../utils/RequestUtils';
 
 const TDStyled = styled.td`
     background-color: ${(props) => props.color};
@@ -26,6 +24,7 @@ const TDStyled = styled.td`
     padding: 5px 0 5px 15px;
     vertical-align: middle;
     word-break: break-all;
+    border-bottom: 1px solid #d5d5d5;
 `;
 
 const TrStyled = styled.tr`
@@ -40,164 +39,57 @@ const TrStyled = styled.tr`
     height: 50px;
 `;
 
-const MenuDiv = styled.div`
-    padding: 5px 10px;
-    background-color: black;
-    color: white;
-    font-size: 12px;
-`;
-
-const LeftSpan = styled.span`
-    padding-left: 20px;
-`;
-
 export default class ReviewRow extends React.Component {
     constructor(props) {
         super(props);
         this.api = this.props.api;
-        this.onSubmitApprove = this.onSubmitApprove.bind(this);
-        this.onReviewSubmit = this.onReviewSubmit.bind(this);
-        this.onClickApproveCancel = this.onClickApproveCancel.bind(this);
-        this.onClickDenyCancel = this.onClickDenyCancel.bind(this);
-        this.onApprove = this.onApprove.bind(this);
-        this.onDeny = this.onDeny.bind(this);
+        this.onReview = this.onReview.bind(this);
         this.state = {
-            name: this.props.details.memberName,
-            showApprove: false,
-            showDeny: false,
+            selectedOption: 'extend',
         };
         this.localDate = new DateUtils();
     }
 
-    onReviewSubmit(message) {
-        this.setState({ reviewMembers: false });
-        if (message) {
-            this.props.onUpdateSuccess(message);
+    onReview(evt) {
+        switch (evt.target.value) {
+            case 'extend':
+                this.props.onUpdate('extend', this.props.details.memberName);
+                break;
+            case 'delete':
+                this.props.onUpdate('delete', this.props.details.memberName);
+                break;
+            case 'no-action':
+                this.props.onUpdate('no-action', this.props.details.memberName);
+                break;
+            default:
+                break;
         }
-    }
-
-    saveJustification(val) {
-        this.setState({ deleteJustification: val });
-    }
-
-    onApprove() {
         this.setState({
-            showApprove: true,
+            selectedOption: evt.target.value,
         });
     }
 
-    onDeny() {
-        this.setState({
-            showDeny: true,
-        });
-    }
-
-    onSubmitApprove() {
-        if (
-            this.props.justificationRequired &&
-            (this.state.deleteJustification === undefined ||
-                this.state.deleteJustification.trim() === '')
-        ) {
+    componentDidUpdate = (prevProps) => {
+        if (prevProps.submittedReview !== this.props.submittedReview) {
             this.setState({
-                errorMessage: 'Justification is required to update a role',
+                selectedOption: 'extend',
             });
-            return;
         }
-
-        let membership = {
-            memberName: this.state.name,
-            approve: true,
-            //TODO: add expiration time in table, is this needed in the table?
-            expiration: '',
-        };
-
-        this.api
-            .processPending(
-                this.props.domain,
-                this.props.role,
-                this.state.name,
-                'Approve pending members in Athenz UI',
-                membership,
-                this.props._csrf
-            )
-            .then(() => {
-                this.setState({
-                    showSuccess: true,
-                    showApprove: false,
-                });
-
-                this.props.onUpdateSuccess();
-            })
-            .catch((err) => {
-                this.setState({
-                    errorMessage: RequestUtils.xhrErrorCheckHelper(err),
-                });
-            });
-    }
-
-    onSubmitDeny() {
-        if (
-            this.props.justificationRequired &&
-            (this.state.deleteJustification === undefined ||
-                this.state.deleteJustification.trim() === '')
-        ) {
-            this.setState({
-                errorMessage: 'Justification is required to update a role',
-            });
-            return;
-        }
-
-        this.api
-            .deletePendingMember(
-                this.props.domain,
-                this.props.role,
-                this.state.name,
-                'Deny pending members in Athenz UI',
-                this.props._csrf
-            )
-            .then(() => {
-                this.setState({
-                    showSuccess: true,
-                    showDeny: false,
-                });
-
-                this.props.onUpdateSuccess();
-            })
-            .catch((err) => {
-                this.setState({
-                    errorMessage: RequestUtils.xhrErrorCheckHelper(err),
-                });
-            });
-    }
-
-    onClickApproveCancel() {
-        this.setState({
-            showApprove: false,
-            errorMessage: '',
-        });
-    }
-
-    onClickDenyCancel() {
-        this.setState({
-            showDeny: false,
-            errorMessage: '',
-        });
-    }
+    };
 
     render() {
+        let rows = [];
         let left = 'left';
         let center = 'center';
         let member = this.props.details;
         let color = this.props.color;
-        let idx = this.props.idx;
+        let exp = member.expiration;
+        if (exp) {
+            exp = this.localDate.getLocalDate(exp, 'UTC', 'UTC');
+        }
 
-        let submitApprove = this.onSubmitApprove.bind(this, true);
-        let submitDeny = this.onSubmitApprove.bind(this, false);
-        let clickApproveCancel = this.onClickApproveCancel.bind(this);
-        let clickDenyCancel = this.onClickDenyCancel.bind(this);
-
-        return (
-            <TrStyled key={this.state.name} data-testid='member-row'>
+        rows.push(
+            <TrStyled key={this.props.idx} data-testid='role-review-row'>
                 <TDStyled color={color} align={left}>
                     {member.memberName}
                 </TDStyled>
@@ -205,68 +97,34 @@ export default class ReviewRow extends React.Component {
                     {member.memberFullName}
                 </TDStyled>
                 <TDStyled color={color} align={left}>
-                    {member.expiration
-                        ? this.localDate.getLocalDate(
-                              member.expiration,
-                              'UTC',
-                              'UTC'
-                          )
-                        : 'N/A'}
+                    {exp}
                 </TDStyled>
                 <TDStyled color={color} align={center}>
                     <RadioButton
                         name={this.props.role + this.props.idx}
-                        value='approve'
-                        checked={false}
-                        onChange={this.onApprove}
+                        value='extend'
+                        checked={this.state.selectedOption === 'extend'}
+                        onChange={this.onReview}
                     />
                 </TDStyled>
                 <TDStyled color={color} align={center}>
                     <RadioButton
                         name={this.props.role + this.props.idx}
-                        value='deny'
-                        checked={false}
-                        onChange={this.onDeny}
+                        value='no-action'
+                        checked={this.state.selectedOption === 'no-action'}
+                        onChange={this.onReview}
                     />
                 </TDStyled>
-                {this.state.showApprove ? (
-                    <UpdateModal
-                        name={this.state.name}
-                        isOpen={this.state.showApprove}
-                        cancel={clickApproveCancel}
-                        submit={submitApprove}
-                        key={this.state.name + '-approve'}
-                        showJustification={this.props.justificationRequired}
-                        message={'Are you sure you want to approve the Member '}
-                        onJustification={this.saveJustification}
-                        errorMessage={this.state.errorMessage}
+                <TDStyled color={color} align={center}>
+                    <RadioButton
+                        name={this.props.role + this.props.idx}
+                        value='delete'
+                        checked={this.state.selectedOption === 'delete'}
+                        onChange={this.onReview}
                     />
-                ) : null}
-                {this.state.showDeny ? (
-                    <UpdateModal
-                        name={this.state.name}
-                        isOpen={this.state.showDeny}
-                        cancel={clickDenyCancel}
-                        submit={submitDeny}
-                        key={this.state.name + '-deny'}
-                        showJustification={this.props.justificationRequired}
-                        message={
-                            'Are you sure you want to deny the approval for Member '
-                        }
-                        onJustification={this.saveJustification}
-                        errorMessage={this.state.errorMessage}
-                    />
-                ) : null}
-                {this.state.showSuccess ? (
-                    <Alert
-                        isOpen={this.state.showSuccess}
-                        title={this.state.successMessage}
-                        onClose={this.closeModal}
-                        type='success'
-                        title='Successfuly update the setting'
-                    />
-                ) : null}
+                </TDStyled>
             </TrStyled>
         );
+        return rows;
     }
 }
