+++ ui/src/pages/role-policy.js
@@ -19,13 +19,15 @@ import UserDomains from '../components/domain/UserDomains';
 import API from '../api';
 import styled from '@emotion/styled';
 import Head from 'next/head';
+import { Link } from '../routes';
 // there is an issue with next-link and next-css if the css is not present then it doesnt load so adding this
 import 'flatpickr/dist/themes/light.css';
-import DomainDetails from '../components/header/DomainDetails';
-import PolicyList from '../components/policy/PolicyList';
+import RoleDetails from '../components/header/RoleDetails';
+import RolePolicyList from '../components/role-policy/RolePolicyList';
 import RequestUtils from '../components/utils/RequestUtils';
-import Tabs from '../components/header/Tabs';
+import RoleTabs from '../components/header/RoleTabs';
 import Error from './_error';
+import NameUtils from '../components/utils/NameUtils';
 
 const AppContainerDiv = styled.div`
     align-items: stretch;
@@ -62,40 +64,50 @@ const TitleDiv = styled.div`
     margin-bottom: 10px;
 `;
 
-export default class DomainPolicyPage extends React.Component {
+const StyledAnchor = styled.a`
+    color: #3570f4;
+    text-decoration: none;
+    cursor: pointer;
+`;
+
+export default class RolePolicyPage extends React.Component {
     static async getInitialProps(props) {
         let api = API(props.req);
         let reload = false;
         let notFound = false;
         let error = undefined;
-        const domains = await Promise.all([
+        const roles = await Promise.all([
             api.listUserDomains(),
             api.getHeaderDetails(),
-            api.getDomain(props.query.domain),
-            api.getPolicies(props.query.domain),
+            api.getRole(
+                props.query.domain,
+                props.query.role,
+                false,
+                false,
+                false
+            ),
+            api.getPolicies(props.query.domain, true),
             api.getForm(),
             api.getPendingDomainRoleMembersList(),
-            api.isAWSTemplateApplied(props.query.domain),
         ]).catch((err) => {
             let response = RequestUtils.errorCheckHelper(err);
             reload = response.reload;
             error = response.error;
             return [{}, {}, {}, {}, {}, {}, {}];
         });
-        let domainDetails = domains[2];
-        domainDetails.isAWSTemplateApplied = !!domains[6];
         return {
             api,
             reload,
             notFound,
             error,
-            domains: domains[0],
-            headerDetails: domains[1],
+            domains: roles[0],
+            headerDetails: roles[1],
             domain: props.query.domain,
-            domainDetails: domainDetails,
-            policies: domains[3],
-            _csrf: domains[4],
-            pending: domains[5],
+            roleDetails: roles[2],
+            role: props.query.role,
+            policies: roles[3],
+            _csrf: roles[4],
+            pending: roles[5],
         };
     }
 
@@ -105,7 +117,34 @@ export default class DomainPolicyPage extends React.Component {
     }
 
     render() {
-        const { domain, reload, domainDetails, policies, _csrf } = this.props;
+        const {
+            domain,
+            role,
+            reload,
+            roleDetails,
+            policies,
+            _csrf,
+        } = this.props;
+
+        let filteredPolicies = policies;
+
+        if (policies) {
+            filteredPolicies = policies.filter((policy) => {
+                let included = false;
+                if (policy.assertions) {
+                    policy.assertions.forEach((element) => {
+                        if (
+                            NameUtils.getShortName(':role.', element.role) ==
+                            role
+                        ) {
+                            included = true;
+                        }
+                    });
+                }
+                return included;
+            });
+        }
+
         if (reload) {
             window.location.reload();
             return <div />;
@@ -128,9 +167,16 @@ export default class DomainPolicyPage extends React.Component {
                         <PoliciesContainerDiv>
                             <PoliciesContentDiv>
                                 <PageHeaderDiv>
-                                    <TitleDiv>{domain}</TitleDiv>
-                                    <DomainDetails
-                                        domainDetails={domainDetails}
+                                    <TitleDiv>
+                                        <Link route='role' params={{ domain }}>
+                                            <StyledAnchor>
+                                                {domain}
+                                            </StyledAnchor>
+                                        </Link>
+                                        / {role}
+                                    </TitleDiv>
+                                    <RoleDetails
+                                        roleDetails={roleDetails}
                                         api={this.api}
                                         _csrf={_csrf}
                                         productMasterLink={
@@ -138,16 +184,18 @@ export default class DomainPolicyPage extends React.Component {
                                                 .productMasterLink
                                         }
                                     />
-                                    <Tabs
+                                    <RoleTabs
                                         api={this.api}
                                         domain={domain}
+                                        role={role}
                                         selectedName={'policies'}
                                     />
                                 </PageHeaderDiv>
-                                <PolicyList
+                                <RolePolicyList
                                     api={this.api}
                                     domain={domain}
-                                    policies={policies}
+                                    role={role}
+                                    policies={filteredPolicies}
                                     _csrf={this.props._csrf}
                                 />
                             </PoliciesContentDiv>
