+++ ui/src/routes.js
@@ -32,7 +32,7 @@ module.exports = routes()
     .add('history', '/domain/:domain/history')
     .add('members', '/domain/:domain/role/:role/members')
     .add('review', '/domain/:domain/role/:role/review')
-    .add('policy', '/domain/:domain/role/:role/policy')
-    .add('domain-policy', '/domain/:domain/policy')
+    .add('role-policy', '/domain/:domain/role/:role/policy')
+    .add('policy', '/domain/:domain/policy')
     .add('settings', '/domain/:domain/role/:role/settings')
     .add('role-history', '/domain/:domain/role/:role/history');
